+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All noteable changes to this project will be documented in this file.
 #### Bug Fixes
 
 - Fixed chatbots besides the default two failing with an error
+- Fixed chatbots without an ending_processor failing
 - Un-commented the "silent_oz" option in the default config.yml
 - Slightly improved startup times